% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{add_anchor_distances}
\alias{add_anchor_distances}
\title{Add anchor distance columns}
\usage{
add_anchor_distances(model_eval, dist_obj, anchor_id, config)
}
\arguments{
\item{model_eval}{Anchor scoring table.}

\item{dist_obj}{Distance object from \code{\link[=build_gower_distances]{build_gower_distances()}}.}

\item{anchor_id}{Anchor model identifier.}
}
\value{
A tibble.
}
\description{
Add anchor distance columns
}
\keyword{internal}
