% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{admissibility_plot_config_data}
\alias{admissibility_plot_config_data}
\title{Resolve raw admissibility plotting config}
\usage{
admissibility_plot_config_data(config = NULL)
}
\arguments{
\item{config}{Raw config list or staged object.}
}
\value{
Named list.
}
\description{
Resolve raw admissibility plotting config
}
\keyword{internal}
