% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{admissibility_plot_trait_label}
\alias{admissibility_plot_trait_label}
\title{Humanize one admissibility trait label}
\usage{
admissibility_plot_trait_label(trait)
}
\arguments{
\item{trait}{Trait code.}
}
\value{
Character scalar.
}
\description{
Humanize one admissibility trait label
}
\keyword{internal}
