% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{admissibility_plot_trait_names}
\alias{admissibility_plot_trait_names}
\title{Resolve one admissibility trait vector}
\usage{
admissibility_plot_trait_names(x)
}
\arguments{
\item{x}{Raw admissibility trait specification.}
}
\value{
Character vector.
}
\description{
Resolve one admissibility trait vector
}
\keyword{internal}
