% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alchemist_s7.R
\name{alchemist_trait_names}
\alias{alchemist_trait_names}
\title{Resolve trait names from a config list element}
\usage{
alchemist_trait_names(traits, models_df)
}
\description{
Resolve trait names from a config list element
}
\keyword{internal}
