% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{apply_cache_defaults}
\alias{apply_cache_defaults}
\title{Apply cache defaults}
\usage{
apply_cache_defaults(config)
}
\arguments{
\item{config}{Config list.}
}
\value{
Config list with cache defaults populated.
}
\description{
Apply cache defaults
}
\keyword{internal}
