% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{apply_workflow_cache_defaults}
\alias{apply_workflow_cache_defaults}
\title{Apply workflow cache defaults}
\usage{
apply_workflow_cache_defaults(config)
}
\arguments{
\item{config}{Workflow-config list.}
}
\value{
Workflow-config list with cache defaults populated.
}
\description{
Apply workflow cache defaults
}
\keyword{internal}
