% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulator_s7.R
\name{as_policy_simulator}
\alias{as_policy_simulator}
\title{Coerce a selector to \code{PolicySimulator}}
\usage{
as_policy_simulator(selector, config = NULL)
}
\arguments{
\item{selector}{A \link{PolicySelector} object.}

\item{config}{Optional simulator config list or \link{Configurer} object.}
}
\value{
A \code{PolicySimulator} object.
}
\description{
Coerce a selector to \code{PolicySimulator}
}
