% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policylearner.R
\name{as_policylearner}
\alias{as_policylearner}
\title{Build a \code{PolicyLearner}}
\usage{
as_policylearner(selector, config = NULL)
}
\arguments{
\item{selector}{A \link{PolicySelector} object.}

\item{config}{Optional learner config list or \link{Configurer} object.}
}
\value{
A \code{PolicyLearner} object.
}
\description{
Build a \code{PolicyLearner}
}
\examples{
\dontrun{
learner <- as_policylearner(selector)
learner
}

}
