% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policyselector.R
\name{as_policyselector}
\alias{as_policyselector}
\title{Build a \code{PolicySelector}}
\usage{
as_policyselector(candidates, config = NULL)
}
\arguments{
\item{candidates}{A \link{Candidates} or \link{Alchemist} object.}

\item{config}{Optional selector config list or \link{Configurer} object.}
}
\value{
A \code{PolicySelector} object.
}
\description{
Build a \code{PolicySelector}
}
\examples{
\dontrun{
selector <- as_policyselector(candidates)
selector
}

}
