% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policysimulator.R
\name{as_policysimulator}
\alias{as_policysimulator}
\title{Build a \code{PolicySimulator}}
\usage{
as_policysimulator(selector, config = NULL)
}
\arguments{
\item{selector}{A \link{PolicySelector} object.}

\item{config}{Optional simulator config list or \link{Configurer} object.}
}
\value{
A \code{PolicySimulator} object.
}
\description{
Build a \code{PolicySimulator}
}
\examples{
\dontrun{
simulator <- as_policysimulator(selector)
simulator
}

}
