% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{assign_ordination_groups}
\alias{assign_ordination_groups}
\title{Assign ordination clusters}
\usage{
assign_ordination_groups(
  points_df,
  max_k = 8,
  min_silhouette = 0.1,
  cluster_col = "nmds_cluster_id"
)
}
\arguments{
\item{points_df}{NMDS point table with \code{MDS1} and \code{MDS2}.}

\item{max_k}{Maximum number of clusters to evaluate.}

\item{min_silhouette}{Minimum acceptable mean silhouette width.}

\item{cluster_col}{Name of the cluster-ID column to create.}
}
\value{
The input point table with cluster columns appended.
}
\description{
Assigns Ward hierarchical clusters to NMDS point coordinates using the best
mean silhouette width up to a maximum \code{k}.
}
