% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{bind_sensitivity_data.PolicySimulator}
\alias{bind_sensitivity_data.PolicySimulator}
\title{Bind scenario benchmark tables from a \link{PolicySimulator}}
\description{
Bind scenario benchmark tables from a \link{PolicySimulator}
}
