% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{build_anchor_audit.PolicyPredictions}
\alias{build_anchor_audit.PolicyPredictions}
\title{Build an anchor support audit from a \link{PolicyPredictions} object}
\description{
Build an anchor support audit from a \link{PolicyPredictions} object
}
