% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{build_anchor_ordination}
\alias{build_anchor_ordination}
\title{Build anchor-specific NMDS info}
\usage{
build_anchor_ordination(
  anchor_row,
  model_scores,
  species_lookup,
  anchor_id_col = "model_id",
  score_id_col = "model_id_chr",
  cluster_col = "nmds_cluster",
  species_col = "species_name"
)
}
\arguments{
\item{anchor_row}{One-row anchor table.}

\item{model_scores}{Model-level NMDS score table.}

\item{species_lookup}{Species lookup returned by \code{\link[=build_species_lookup]{build_species_lookup()}}.}

\item{anchor_id_col}{Anchor-ID column name.}

\item{score_id_col}{Model-score ID column name.}

\item{cluster_col}{Cluster column name in \code{model_scores}.}

\item{species_col}{Species-name column name.}
}
\value{
A list.
}
\description{
Build anchor-specific NMDS info
}
