% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjurer_s7.R
\name{build_conjurer}
\alias{build_conjurer}
\title{Build a \code{Conjurer}}
\usage{
build_conjurer(selector, learner = NULL, config = NULL)
}
\arguments{
\item{selector}{A \link{PolicySelector} object.}

\item{learner}{Optional fitted \link{PolicyLearner}.}

\item{config}{Optional conjurer config list or \link{Configurer} object.}
}
\value{
A \code{Conjurer} object.
}
\description{
Build a \code{Conjurer}
}
\examples{
\dontrun{
conjurer <- as_conjurer(selector)
conjurer
}

}
