% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{build_constructed_policy_definition}
\alias{build_constructed_policy_definition}
\title{Build one constructed policy definition}
\usage{
build_constructed_policy_definition(
  group_def,
  metric_def,
  constructor_def = list()
)
}
\arguments{
\item{group_def}{One normalized policy-group definition.}

\item{metric_def}{One normalized policy-metric definition.}

\item{constructor_def}{Optional constructor overrides.}
}
\value{
One policy-definition list.
}
\description{
Build one constructed policy definition
}
\keyword{internal}
