% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selection.R
\name{build_equivalence_sets}
\alias{build_equivalence_sets}
\title{Build policy-equivalence classes}
\usage{
build_equivalence_sets(select_ref, pair_tbl)
}
\arguments{
\item{select_ref}{Global policy-selection reference table.}

\item{pair_tbl}{Pairwise policy-equivalence table.}
}
\value{
A tibble.
}
\description{
Converts pairwise equivalence calls to connected equivalence classes.
}
