% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{build_ordination_hulls}
\alias{build_ordination_hulls}
\title{Build NMDS cluster hulls}
\usage{
build_ordination_hulls(
  points_df,
  cluster_col = "nmds_cluster_id",
  min_points = 3L
)
}
\arguments{
\item{points_df}{Clustered NMDS point table.}

\item{cluster_col}{Cluster-ID column name.}

\item{min_points}{Minimum number of points required to form a hull.}
}
\value{
A tibble.
}
\description{
Build NMDS cluster hulls
}
