% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{build_policy_definition_from_name}
\alias{build_policy_definition_from_name}
\title{Build one policy definition from its canonical name}
\usage{
build_policy_definition_from_name(
  policy_name,
  policy_path = NULL,
  registry = NULL
)
}
\arguments{
\item{policy_name}{Canonical policy name.}

\item{policy_path}{Optional registry path.}

\item{registry}{Optional parsed policy registry.}
}
\value{
One policy-definition list or \code{NULL}.
}
\description{
Build one policy definition from its canonical name
}
\keyword{internal}
