% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulator_s7.R
\name{build_policy_simulator}
\alias{build_policy_simulator}
\title{Build a \code{PolicySimulator}}
\usage{
build_policy_simulator(selector, config = NULL)
}
\arguments{
\item{selector}{A \link{PolicySelector} object.}

\item{config}{Optional simulator config list or \link{Configurer} object.}
}
\value{
A \code{PolicySimulator} object.
}
\description{
Build a \code{PolicySimulator}
}
\examples{
\dontrun{
simulator <- as_policy_simulator(selector)
simulator
}

}
