% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{build_sensitivity_table.PolicySimulator}
\alias{build_sensitivity_table.PolicySimulator}
\title{Build a scenario manifest from a \link{PolicySimulator}}
\description{
Build a scenario manifest from a \link{PolicySimulator}
}
