% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{calculate_range_overlap_vec}
\alias{calculate_range_overlap_vec}
\title{Compute vectorized directional interval overlap}
\usage{
calculate_range_overlap_vec(a_min, a_max, b_min, b_max)
}
\arguments{
\item{a_min}{Anchor minimum scalar.}

\item{a_max}{Anchor maximum scalar.}

\item{b_min}{Candidate minimum vector.}

\item{b_max}{Candidate maximum vector.}
}
\value{
Numeric vector.
}
\description{
Compute vectorized directional interval overlap
}
\keyword{internal}
