% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{calibrate_uncertainty}
\alias{calibrate_uncertainty}
\title{Calibrate uncertainty on a staged staged object}
\usage{
calibrate_uncertainty(object, ...)
}
\arguments{
\item{object}{A staged staged object such as a \link{PolicySelector} or
\link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
An updated staged staged object.
}
\description{
Dispatches to the appropriate uncertainty-calibration method for the
supplied staged staged object.
}
