% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates_s7.R
\name{candidate_refresh_value}
\alias{candidate_refresh_value}
\title{Validate a boolean refresh-like field}
\usage{
candidate_refresh_value(x, field_name)
}
\arguments{
\item{x}{Value to validate.}

\item{field_name}{Field label used in error messages.}
}
\value{
Logical scalar.
}
\description{
Validate a boolean refresh-like field
}
\keyword{internal}
