% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates_s7.R
\name{candidates_similarity_key_cols}
\alias{candidates_similarity_key_cols}
\title{Test whether an object is a \code{Candidates} instance}
\usage{
candidates_similarity_key_cols(candidates)
}
\arguments{
\item{candidates}{A \link{Candidates} object.}
}
\value{
Character vector.
}
\description{
Resolve the key similarity trait columns stored on \code{Candidates}
}
\keyword{internal}
