% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{canonical_policy_name}
\alias{canonical_policy_name}
\title{Canonicalize policy identifiers}
\usage{
canonical_policy_name(policy_name, policy_path = NULL)
}
\arguments{
\item{policy_name}{Character vector of policy identifiers.}

\item{policy_path}{Optional path to a policy registry JSON file.}
}
\value{
Character vector of canonical coded names.
}
\description{
Canonicalize policy identifiers
}
\keyword{internal}
