% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{collect_sensitivity_results.PolicySimulator}
\alias{collect_sensitivity_results.PolicySimulator}
\title{Collect scenario benchmark tables from a \link{PolicySimulator}}
\description{
Collect scenario benchmark tables from a \link{PolicySimulator}
}
