% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_s7.R
\name{collect_sentinel_results}
\alias{collect_sentinel_results}
\title{Collect Sentinel summary outputs from disk}
\usage{
collect_sentinel_results(object, completed_only = TRUE)
}
\arguments{
\item{object}{A \link{Sentinel} object.}

\item{completed_only}{Logical scalar.}
}
\value{
Tibble.
}
\description{
Collect Sentinel summary outputs from disk
}
