% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{command_line_true}
\alias{command_line_true}
\title{Parse a command-line boolean}
\usage{
command_line_true(value)
}
\arguments{
\item{value}{Value to parse.}
}
\value{
Logical scalar.
}
\description{
Converts common command-line boolean strings to \code{TRUE} or \code{FALSE}.
}
\keyword{internal}
