% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compute_frequency_span}
\alias{compute_frequency_span}
\title{Compute the log-frequency span}
\usage{
compute_frequency_span(frequency)
}
\arguments{
\item{frequency}{Numeric or coercible frequency vector.}
}
\value{
Numeric scalar.
}
\description{
Compute the log-frequency span
}
\keyword{internal}
