% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_data}
\alias{config_data}
\title{Resolve config data}
\usage{
config_data(config)
}
\arguments{
\item{config}{Config input.}
}
\value{
A config list.
}
\description{
Resolve config data
}
\keyword{internal}
