% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_option_values}
\alias{config_option_values}
\title{Build config option values}
\usage{
config_option_values(config)
}
\arguments{
\item{config}{Normalized config list.}
}
\value{
Named list of option values.
}
\description{
Returns a named list of options suitable for \code{options(...)}.
}
