% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{conjurer_metric_label}
\alias{conjurer_metric_label}
\title{Resolve one default label for a \code{Conjurer} metric}
\usage{
conjurer_metric_label(metric)
}
\arguments{
\item{metric}{Metric name.}
}
\value{
Character scalar.
}
\description{
Resolve one default label for a \code{Conjurer} metric
}
\keyword{internal}
