% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{conjurer_metric_value_labels}
\alias{conjurer_metric_value_labels}
\title{Format \code{Conjurer} heatmap values}
\usage{
conjurer_metric_value_labels(values, metric)
}
\arguments{
\item{values}{Numeric vector.}

\item{metric}{Metric name.}
}
\value{
Character vector.
}
\description{
Format \code{Conjurer} heatmap values
}
\keyword{internal}
