% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjurer_s7.R
\name{conjurer_summarize_draws}
\alias{conjurer_summarize_draws}
\title{Summarize one set of conjurer draws}
\usage{
conjurer_summarize_draws(selected_draws, consensus_draws, baseline_selected)
}
\arguments{
\item{selected_draws}{Draw-level selected-policy table.}

\item{consensus_draws}{Draw-level donor-consensus table.}

\item{baseline_selected}{Baseline selected-policy table.}
}
\value{
A tibble.
}
\description{
Summarize one set of conjurer draws
}
\keyword{internal}
