% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{conjurer_trait_labels}
\alias{conjurer_trait_labels}
\title{Resolve one default trait-label map for \code{Conjurer}}
\usage{
conjurer_trait_labels(traits)
}
\arguments{
\item{traits}{Character vector.}
}
\value{
Named character vector.
}
\description{
Resolve one default trait-label map for \code{Conjurer}
}
\keyword{internal}
