% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{construct_sensitivity_table.PolicySimulator}
\alias{construct_sensitivity_table.PolicySimulator}
\title{Construct a scenario manifest from a \link{PolicySimulator}}
\description{
Construct a scenario manifest from a \link{PolicySimulator}
}
