% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{create_policy_selector}
\alias{create_policy_selector}
\title{Build a \code{PolicySelector}}
\usage{
create_policy_selector(candidates, config = NULL)
}
\arguments{
\item{candidates}{A \link{Candidates} object.}

\item{config}{Optional selector config list or \link{Configurer} object.}
}
\value{
A \code{PolicySelector} object.
}
\description{
Build a \code{PolicySelector}
}
\examples{
\dontrun{
selector <- create_policy_selector(candidates)
selector
}

}
