% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_policy.R
\name{default_meta_policy_exclusions}
\alias{default_meta_policy_exclusions}
\title{Default meta-policy feature exclusions}
\usage{
default_meta_policy_exclusions()
}
\value{
Character vector of non-feature columns that should never be used as
meta-policy predictors.
}
\description{
Default meta-policy feature exclusions
}
\keyword{internal}
