% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{empty_scorecard}
\alias{empty_scorecard}
\title{Create an empty \code{Scorecard}}
\usage{
empty_scorecard()
}
\value{
A \code{Scorecard} object with empty tables.
}
\description{
Create an empty \code{Scorecard}
}
\keyword{internal}
