% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_parent_path}
\alias{ensure_parent_path}
\title{Create a parent directory}
\usage{
ensure_parent_path(path)
}
\arguments{
\item{path}{Character scalar file path.}
}
\value{
The input path, invisibly.
}
\description{
Creates the parent directory for a file path when it does not already exist.
}
\keyword{internal}
