% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{equal_start_weights}
\alias{equal_start_weights}
\title{Replace one trait-weight map with equal starts}
\usage{
equal_start_weights(trait_map)
}
\arguments{
\item{trait_map}{Named numeric vector or list.}
}
\value{
Named numeric vector with all weights set to \code{1}.
}
\description{
Replace one trait-weight map with equal starts
}
\keyword{internal}
