% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{equation_multiplier}
\alias{equation_multiplier}
\title{Compute biomass multiplier for an equation}
\usage{
equation_multiplier(
  slope,
  intercept,
  anchor_sigma,
  anchor_pdf,
  anchor_pdf_precomp = NULL
)
}
\arguments{
\item{slope}{Numeric slope.}

\item{intercept}{Numeric intercept.}

\item{anchor_sigma}{Anchor mean backscatter.}

\item{anchor_pdf}{Anchor length-density table.}
}
\value{
One-row tibble.
}
\description{
Compute biomass multiplier for an equation
}
\keyword{internal}
