% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{equation_sigma_mean}
\alias{equation_sigma_mean}
\title{Compute mean backscatter for an equation}
\usage{
equation_sigma_mean(slope, intercept, anchor_pdf, anchor_pdf_precomp = NULL)
}
\arguments{
\item{slope}{Numeric slope.}

\item{intercept}{Numeric intercept.}

\item{anchor_pdf}{Anchor length-density table.}
}
\value{
Numeric scalar.
}
\description{
Compute mean backscatter for an equation
}
\keyword{internal}
