% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{group_model_rows}
\alias{group_model_rows}
\title{Identify generalized-model rows}
\usage{
group_model_rows(rows)
}
\arguments{
\item{rows}{Candidate-policy row table.}
}
\value{
Tibble.
}
\description{
Identify generalized-model rows
}
\keyword{internal}
