% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installed_template_path}
\alias{installed_template_path}
\title{Resolve an installed template path}
\usage{
installed_template_path(name)
}
\arguments{
\item{name}{Template filename under \code{inst/templates}.}
}
\value{
Character scalar path.
}
\description{
Resolve an installed template path
}
\keyword{internal}
