% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_policy.R
\name{is_default_meta_policy_feature}
\alias{is_default_meta_policy_feature}
\title{Test whether one column is eligible as a default meta-policy feature}
\usage{
is_default_meta_policy_feature(name, x)
}
\arguments{
\item{name}{Column name.}

\item{x}{Column vector.}
}
\value{
Logical scalar.
}
\description{
Test whether one column is eligible as a default meta-policy feature
}
\keyword{internal}
