% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{legacy_config_names}
\alias{legacy_config_names}
\title{Return disallowed legacy config names}
\usage{
legacy_config_names()
}
\value{
Named list of legacy-to-current field-name mappings by section.
}
\description{
Return disallowed legacy config names
}
\keyword{internal}
