% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_cfg}
\alias{merge_cfg}
\title{Recursively merge two lists}
\usage{
merge_cfg(x, y)
}
\arguments{
\item{x}{Base list.}

\item{y}{Override list.}
}
\value{
A recursively merged list.
}
\description{
Recursively merge two lists
}
\keyword{internal}
