% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_active_policy_names}
\alias{normalize_active_policy_names}
\title{Normalize active policy names}
\usage{
normalize_active_policy_names(config)
}
\arguments{
\item{config}{Workflow-config list.}
}
\value{
Workflow-config list with canonical active policy names.
}
\description{
Normalize active policy names
}
\keyword{internal}
