% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_frequency_mode}
\alias{normalize_frequency_mode}
\title{Normalize one frequency mode}
\usage{
normalize_frequency_mode(mode)
}
\arguments{
\item{mode}{Frequency coherence mode.}
}
\value{
Character scalar.
}
\description{
Normalize one frequency mode
}
\keyword{internal}
