% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Return a fallback value for \code{NULL} or empty input}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Primary object.}

\item{y}{Fallback object.}
}
\value{
\code{x} when present, otherwise \code{y}.
}
\description{
Return a fallback value for \code{NULL} or empty input
}
\keyword{internal}
