% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_anchor_summary}
\alias{plot_anchor_summary}
\title{Plot integrated anchor summary}
\usage{
plot_anchor_summary(integrated_tbl, score_tbl, interval_tbl)
}
\arguments{
\item{integrated_tbl}{Integrated anchor-summary table.}

\item{score_tbl}{Admissible candidate-score table.}

\item{interval_tbl}{All-policy interval table.}
}
\value{
A ggplot object.
}
\description{
Plot integrated anchor summary
}
