% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_biomass_candidate_map}
\alias{plot_biomass_candidate_map}
\title{Plot candidate biomass response}
\usage{
plot_biomass_candidate_map(candidate_tbl, anchor_label)
}
\arguments{
\item{candidate_tbl}{Candidate-biomass table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot candidate biomass response
}
