% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_component_importance}
\alias{plot_component_importance}
\title{Plot tuning component importance}
\usage{
plot_component_importance(impact_tbl, label_map = NULL)
}
\arguments{
\item{impact_tbl}{Tuning component-impact table.}

\item{label_map}{Optional named vector mapping component codes to labels.}
}
\value{
A ggplot object.
}
\description{
Plot tuning component importance
}
