% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_conformal_scores}
\alias{plot_conformal_scores}
\title{Plot conformal calibration by policy}
\usage{
plot_conformal_scores(cal_tbl)
}
\arguments{
\item{cal_tbl}{Policy-level conformal calibration table.}
}
\value{
A ggplot object.
}
\description{
Plot conformal calibration by policy
}
