% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_interval_panel}
\alias{plot_interval_panel}
\title{Plot all-reference policy interval panel}
\usage{
plot_interval_panel(interval_tbl)
}
\arguments{
\item{interval_tbl}{All-policy interval table across references.}
}
\value{
A ggplot object.
}
\description{
Plot all-reference policy interval panel
}
