% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_length_density}
\alias{plot_length_density}
\title{Plot a length-density curve}
\usage{
plot_length_density(length_tbl, anchor_label)
}
\arguments{
\item{length_tbl}{Length-density table with \code{length_cm} and \code{f_len}.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot a length-density curve
}
