% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_length_density_panel}
\alias{plot_length_density_panel}
\title{Plot a panel of anchor length-density curves}
\usage{
plot_length_density_panel(length_tbl, reference_col = "anchor_species")
}
\arguments{
\item{length_tbl}{Bound anchor length-density table.}

\item{reference_col}{Facet-label column.}
}
\value{
A ggplot object.
}
\description{
Plot a panel of anchor length-density curves
}
