% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_model_weights}
\alias{plot_model_weights}
\title{Plot model weights against distance}
\usage{
plot_model_weights(weight_tbl, anchor_label)
}
\arguments{
\item{weight_tbl}{Candidate-weight table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot model weights against distance
}
