% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_multiplier_vs_expected_length}
\alias{plot_multiplier_vs_expected_length}
\title{Plot selected biomass multiplier against expected anchor length}
\usage{
plot_multiplier_vs_expected_length(anchor_tbl)
}
\arguments{
\item{anchor_tbl}{Selected-policy table with multiplier intervals and expected length.}
}
\value{
A ggplot object.
}
\description{
Plot selected biomass multiplier against expected anchor length
}
