% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ordination_clusters}
\alias{plot_ordination_clusters}
\title{Plot NMDS clusters}
\usage{
plot_ordination_clusters(
  points_tbl,
  cluster_col = "policy_cluster_id",
  reference_col = "is_reference",
  species_col = "species_name",
  common_col = "common",
  colorbar_name = "NMDS Cluster ID"
)
}
\arguments{
\item{points_tbl}{NMDS point table.}

\item{cluster_col}{Cluster-label column.}

\item{reference_col}{Reference-flag column.}

\item{species_col}{Species-label column.}

\item{common_col}{Optional common-name column.}

\item{colorbar_name}{Legend title for the cluster color scale.}
}
\value{
A ggplot object.
}
\description{
Plot NMDS clusters
}
