% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ordination_vectors}
\alias{plot_ordination_vectors}
\title{Plot NMDS variable vectors}
\usage{
plot_ordination_vectors(
  vec_tbl,
  points_tbl,
  reference_col = "is_reference",
  species_col = "species_name",
  common_col = "common"
)
}
\arguments{
\item{vec_tbl}{NMDS vector table.}

\item{points_tbl}{NMDS point table.}

\item{reference_col}{Reference-flag column.}

\item{species_col}{Species-label column.}

\item{common_col}{Optional common-name column.}
}
\value{
A ggplot object.
}
\description{
Plot NMDS variable vectors
}
