% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_overlap_heatmap}
\alias{plot_overlap_heatmap}
\title{Plot overlap heatmap}
\usage{
plot_overlap_heatmap(overlap_tbl, metric_labs = NULL)
}
\arguments{
\item{overlap_tbl}{Overlap-summary table.}

\item{metric_labs}{Optional named vector mapping metric codes to labels.}
}
\value{
A ggplot object.
}
\description{
Plot overlap heatmap
}
