% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pairwise_pivot_histogram}
\alias{plot_pairwise_pivot_histogram}
\title{Plot pairwise pivot distribution}
\usage{
plot_pairwise_pivot_histogram(pairwise_tbl, summary_tbl, anchor_label)
}
\arguments{
\item{pairwise_tbl}{Pairwise pivot table.}

\item{summary_tbl}{One-row pivot summary table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot pairwise pivot distribution
}
