% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_policy_boxplot}
\alias{plot_policy_boxplot}
\title{Plot pseudo-anchor policy errors}
\usage{
plot_policy_boxplot(perf_tbl)
}
\arguments{
\item{perf_tbl}{Pseudo-anchor benchmark table.}
}
\value{
A ggplot object.
}
\description{
Plot pseudo-anchor policy errors
}
