% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_policy_coefficients}
\alias{plot_policy_coefficients}
\title{Plot selected-policy coefficient intervals}
\usage{
plot_policy_coefficients(coefficient_tbl)
}
\arguments{
\item{coefficient_tbl}{Selected-policy table with coefficient interval columns.}
}
\value{
A ggplot object.
}
\description{
Plot selected-policy coefficient intervals
}
