% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_policy_heatmap}
\alias{plot_policy_heatmap}
\title{Plot species-blocked policy heatmap}
\usage{
plot_policy_heatmap(perf_tbl, policy_labs = NULL)
}
\arguments{
\item{perf_tbl}{Species-block benchmark table.}

\item{policy_labs}{Optional named vector mapping policy codes to labels.}
}
\value{
A ggplot object.
}
\description{
Plot species-blocked policy heatmap
}
