% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_selected_intervals}
\alias{plot_selected_intervals}
\title{Plot selected policy intervals}
\usage{
plot_selected_intervals(sel_tbl, reference_label = "Reference")
}
\arguments{
\item{sel_tbl}{Selected-policy interval table.}

\item{reference_label}{Reference label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot selected policy intervals
}
