% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_selected_policy_species_rank}
\alias{plot_selected_policy_species_rank}
\title{Plot selected policies against species-block oracle ranks}
\usage{
plot_selected_policy_species_rank(diagnostics_tbl)
}
\arguments{
\item{diagnostics_tbl}{Output from
\code{\link[=compare_selected_policy_species_rank]{compare_selected_policy_species_rank()}}.}
}
\value{
A ggplot object.
}
\description{
Plot selected policies against species-block oracle ranks
}
