% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_sensitivity_overview}
\alias{plot_sensitivity_overview}
\title{Plot sensitivity summary}
\usage{
plot_sensitivity_overview(plot_tbl)
}
\arguments{
\item{plot_tbl}{Long sensitivity-summary table.}
}
\value{
A ggplot object.
}
\description{
Plot sensitivity summary
}
