% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_similarity_map}
\alias{plot_similarity_map}
\title{Plot the admissible similarity map}
\usage{
plot_similarity_map(map_tbl, anchor_label)
}
\arguments{
\item{map_tbl}{Admissible candidate table with distance and weight columns.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot the admissible similarity map
}
