% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_slope_distribution}
\alias{plot_slope_distribution}
\title{Plot the overall slope distribution}
\usage{
plot_slope_distribution(slope_tbl)
}
\arguments{
\item{slope_tbl}{Study-cell slope-summary table.}
}
\value{
A ggplot object.
}
\description{
Plot the overall slope distribution
}
