% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_slope_group}
\alias{plot_slope_group}
\title{Plot slope distributions by group}
\usage{
plot_slope_group(slope_tbl)
}
\arguments{
\item{slope_tbl}{Study-cell slope-summary table.}
}
\value{
A ggplot object.
}
\description{
Plot slope distributions by group
}
