% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_slope_support}
\alias{plot_slope_support}
\title{Plot slope-support proportions by group}
\usage{
plot_slope_support(support_tbl)
}
\arguments{
\item{support_tbl}{Weighted slope-support table.}
}
\value{
A ggplot object.
}
\description{
Plot slope-support proportions by group
}
