% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_species_boxplot}
\alias{plot_species_boxplot}
\title{Plot species-blocked policy errors}
\usage{
plot_species_boxplot(perf_tbl)
}
\arguments{
\item{perf_tbl}{Species-block benchmark table.}
}
\value{
A ggplot object.
}
\description{
Plot species-blocked policy errors
}
