% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_species_ordination}
\alias{plot_species_ordination}
\title{Plot species-level NMDS ordination}
\usage{
plot_species_ordination(
  points_tbl,
  vec_tbl = NULL,
  fac_tbl = NULL,
  reference_col = "is_reference",
  species_col = "species_name",
  group_col = "species_cluster_id",
  colorbar_name = "Cluster ID & ellipse",
  ellipse_level = NULL
)
}
\arguments{
\item{points_tbl}{Species-level NMDS point table.}

\item{vec_tbl}{Optional NMDS vector table.}

\item{fac_tbl}{Optional NMDS factor-centroid table.}

\item{reference_col}{Reference-flag column.}

\item{species_col}{Species-label column.}

\item{group_col}{Grouping/fill column.}

\item{colorbar_name}{Colorbar title.}

\item{ellipse_level}{Optional ellipse interval bounded by 0 and 1.}
}
\value{
A ggplot object.
}
\description{
Plot species-level NMDS ordination
}
