% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_species_policy_ranked}
\alias{plot_species_policy_ranked}
\title{Plot every species-policy benchmark rank}
\usage{
plot_species_policy_ranked(perf_tbl)
}
\arguments{
\item{perf_tbl}{Species-block benchmark table or output from
\code{\link[=summarize_species_policy_performance]{summarize_species_policy_performance()}}.}
}
\value{
A ggplot object.
}
\description{
Plot every species-policy benchmark rank
}
