% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_top_models}
\alias{plot_top_models}
\title{Plot top candidate weights}
\usage{
plot_top_models(top_tbl, anchor_label)
}
\arguments{
\item{top_tbl}{Ranked top-candidate table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot top candidate weights
}
