% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ts_ribbon}
\alias{plot_ts_ribbon}
\title{Plot a weighted TS ribbon}
\usage{
plot_ts_ribbon(ribbon_tbl, anchor_label)
}
\arguments{
\item{ribbon_tbl}{Weighted TS summary table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot a weighted TS ribbon
}
