% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_tuning_variation}
\alias{plot_tuning_variation}
\title{Plot tuning resample variability}
\usage{
plot_tuning_variation(plot_tbl, block_col = "block")
}
\arguments{
\item{plot_tbl}{Tuning-resample block-summary table.}

\item{block_col}{Block-name column.}
}
\value{
A ggplot object.
}
\description{
Plot tuning resample variability
}
