% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_uncertainty_blocks}
\alias{plot_uncertainty_blocks}
\title{Plot uncertainty blocks}
\usage{
plot_uncertainty_blocks(dropout_tbl, anchor_label)
}
\arguments{
\item{dropout_tbl}{Uncertainty dropout table.}

\item{anchor_label}{Anchor label used in the title.}
}
\value{
A ggplot object.
}
\description{
Plot uncertainty blocks
}
