% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_uncertainty_heat}
\alias{plot_uncertainty_heat}
\title{Plot uncertainty heatmap}
\usage{
plot_uncertainty_heat(dropout_tbl, block_labs = NULL)
}
\arguments{
\item{dropout_tbl}{Anchor-level dropout summary table.}

\item{block_labs}{Optional named vector mapping block codes to labels.}
}
\value{
A ggplot object.
}
\description{
Plot uncertainty heatmap
}
