% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{policy_equation_row}
\alias{policy_equation_row}
\title{Build an equation row}
\usage{
policy_equation_row(slope, intercept)
}
\arguments{
\item{slope}{Numeric slope.}

\item{intercept}{Numeric intercept.}
}
\value{
One-row tibble.
}
\description{
Build an equation row
}
\keyword{internal}
