% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{policy_selector_config_data}
\alias{policy_selector_config_data}
\title{Test whether an object is a \code{PolicySelector} instance}
\usage{
policy_selector_config_data(config)
}
\arguments{
\item{config}{Optional config input.}
}
\value{
A list.
}
\description{
Normalize policy-selector config input
}
\keyword{internal}
