% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_selection_conformal.R
\name{post_selection_quantile}
\alias{post_selection_quantile}
\title{Finite-sample post-selection conformal quantile}
\usage{
post_selection_quantile(x, alpha)
}
\arguments{
\item{x}{Numeric residual vector.}

\item{alpha}{Miscoverage level.}
}
\value{
Numeric scalar.
}
\description{
Finite-sample post-selection conformal quantile
}
\keyword{internal}
