% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_policy.R
\name{predict_meta_policy_score}
\alias{predict_meta_policy_score}
\title{Predict policy-specific transferability score with a meta-policy learner}
\usage{
predict_meta_policy_score(object, new_policy_tbl)
}
\arguments{
\item{object}{Fitted object from \code{fit_meta_policy_learner()}.}

\item{new_policy_tbl}{New policy table.}
}
\value{
\code{new_policy_tbl} with \code{.meta_predicted_score}.
}
\description{
Predict policy-specific transferability score with a meta-policy learner
}
\examples{
\dontrun{
scored <- predict_meta_policy_score(
  learner,
  new_policy_tbl = selector@predictions@intervals
)
}
}
