% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates_s7.R
\name{print.Candidates}
\alias{print.Candidates}
\title{Print a \code{Candidates}}
\arguments{
\item{x}{A \link{Candidates} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{Candidates}
}
\keyword{internal}
