% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurer_s7.R
\name{print.Configurer}
\alias{print.Configurer}
\title{Print a \code{Configurer}}
\arguments{
\item{x}{A \link{Configurer} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{Configurer}
}
\keyword{internal}
