% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjurer_s7.R
\name{print.Conjurer}
\alias{print.Conjurer}
\title{Print a \code{Conjurer}}
\arguments{
\item{x}{A \link{Conjurer} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{Conjurer}
}
\keyword{internal}
