% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{print.PolicyPredictions}
\alias{print.PolicyPredictions}
\title{Print a \code{PolicyPredictions}}
\arguments{
\item{x}{A \link{PolicyPredictions} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{PolicyPredictions}
}
\keyword{internal}
