% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{print.PolicySelector}
\alias{print.PolicySelector}
\title{Print a \code{PolicySelector}}
\arguments{
\item{x}{A \link{PolicySelector} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{PolicySelector}
}
\keyword{internal}
