% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulator_s7.R
\name{print.PolicySimulator}
\alias{print.PolicySimulator}
\title{Print a \code{PolicySimulator}}
\arguments{
\item{x}{A \link{PolicySimulator} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{PolicySimulator}
}
\keyword{internal}
