% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{print.Referee}
\alias{print.Referee}
\title{Print a \code{Referee}}
\arguments{
\item{x}{A \link{Referee} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{Referee}
}
\keyword{internal}
