% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{print.Scorecard}
\alias{print.Scorecard}
\title{Print a \code{Scorecard}}
\arguments{
\item{x}{A \link{Scorecard} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{Scorecard}
}
\keyword{internal}
