% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_config_s7.R
\name{print.WorkflowConfig}
\alias{print.WorkflowConfig}
\title{Print a \code{WorkflowConfig}}
\arguments{
\item{x}{A \link{WorkflowConfig} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{WorkflowConfig}
}
\keyword{internal}
