% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{reference_length}
\alias{reference_length}
\title{Evaluate the reference length for one model row}
\usage{
reference_length(row_df)
}
\arguments{
\item{row_df}{One-row data frame.}
}
\value{
Numeric scalar.
}
\description{
Evaluate the reference length for one model row
}
\keyword{internal}
