% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{reject_legacy_config_names}
\alias{reject_legacy_config_names}
\title{Reject legacy config names}
\usage{
reject_legacy_config_names(config)
}
\arguments{
\item{config}{Raw config list.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Reject legacy config names
}
\keyword{internal}
