% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_progress}
\alias{report_progress}
\title{Emit one progress message}
\usage{
report_progress(progress, ...)
}
\arguments{
\item{progress}{Logical scalar.}

\item{...}{Message fragments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Emit one progress message
}
\keyword{internal}
