% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_equivalent_policy_sets}
\alias{resolve_equivalent_policy_sets}
\title{Resolve equivalent policy sets from a table}
\usage{
resolve_equivalent_policy_sets(tbl)
}
\arguments{
\item{tbl}{Data frame or tibble.}
}
\value{
Character vector.
}
\description{
Resolve equivalent policy sets from a table
}
\keyword{internal}
