% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_selected_policy_names}
\alias{resolve_selected_policy_names}
\title{Resolve displayed selected policy names from a table}
\usage{
resolve_selected_policy_names(tbl)
}
\arguments{
\item{tbl}{Data frame or tibble.}
}
\value{
Character vector.
}
\description{
Resolve displayed selected policy names from a table
}
\keyword{internal}
