% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{resolve_similarity_setup}
\alias{resolve_similarity_setup}
\title{Normalize similarity config options}
\usage{
resolve_similarity_setup(cfg_user, alpha, k_species, k_study)
}
\arguments{
\item{cfg_user}{Normalized user config list.}

\item{alpha}{Starting alpha value.}

\item{k_species}{Starting species-kernel value.}

\item{k_study}{Starting study-kernel value.}
}
\value{
A normalized config list.
}
\description{
Normalize similarity config options
}
\keyword{internal}
