% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_s7.R
\name{resume_sentinel}
\alias{resume_sentinel}
\title{Resume a Sentinel run from disk}
\usage{
resume_sentinel(object)
}
\arguments{
\item{object}{A \link{Sentinel} object.}
}
\value{
An updated \link{Sentinel} object.
}
\description{
Resume a Sentinel run from disk
}
