% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{run_script_from_config}
\alias{run_script_from_config}
\title{Run the packaged script}
\usage{
run_script_from_config(
  config_path,
  script_name = "swfscfish.R",
  rscript_path = NULL,
  wait = TRUE
)
}
\arguments{
\item{config_path}{Config YAML path.}

\item{script_name}{Packaged script name.}

\item{rscript_path}{Optional \code{Rscript} executable path.}

\item{wait}{Logical scalar. If \code{TRUE}, wait for the script to finish.}
}
\value{
The \code{system2()} exit status.
}
\description{
Launches the packaged script wrapper with a validated config YAML file.
}
