% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{score_tasks_parallel}
\alias{score_tasks_parallel}
\title{Run similarity tasks in parallel}
\usage{
score_tasks_parallel(tasks, worker_name, workers, shared_data, warning_text)
}
\arguments{
\item{tasks}{Task list.}

\item{worker_name}{Internal worker-function name.}

\item{workers}{Number of workers.}

\item{shared_data}{Shared worker data exported once per cluster.}

\item{warning_text}{Warning prefix used if the parallel path fails.}
}
\value{
A list of task results.
}
\description{
Run similarity tasks in parallel
}
\keyword{internal}
