% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{Scorecard-class}
\alias{Scorecard-class}
\alias{Scorecard}
\title{Typed recommendation scorecard}
\usage{
Scorecard(
  intervals = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  selected = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  ts_panel = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  recommendation_cards = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  surrogate_rules = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  consensus = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  anchor_summary = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  anchor_audit = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  species_coverage = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  selection_diagnostics = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  key_missing_overall = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  key_missing_by_field = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  key_missing_by_model = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  anchor_missing_gate = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  status = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })()
)
}
\description{
Stores the post-prediction recommendation tables and diagnostics built by
\code{\link[=predict]{predict()}} on a \link{Referee}.
}
\examples{
\dontrun{
scorecard <- predict(as_referee(selector, predictions = predictions))
scorecard@anchor_summary
scorecard@selection_diagnostics
}

}
