% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{screen_missing_metadata.Candidates}
\alias{screen_missing_metadata.Candidates}
\title{Add key-metadata missingness for a \link{Candidates} object}
\description{
Add key-metadata missingness for a \link{Candidates} object
}
\keyword{internal}
