% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_s7.R
\name{sentinel_validate_prediction_table}
\alias{sentinel_validate_prediction_table}
\title{Validate one Sentinel prediction table}
\usage{
sentinel_validate_prediction_table(tbl, field_name)
}
\arguments{
\item{tbl}{Prediction table supplied by a user workflow.}

\item{field_name}{User-facing field label used in error messages.}
}
\value{
A tibble.
}
\description{
Validate one Sentinel prediction table
}
\keyword{internal}
