% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{set_reference_length_pdf}
\alias{set_reference_length_pdf}
\title{Set user-supplied reference length PDFs}
\usage{
set_reference_length_pdf(object, length_pdf)
}
\arguments{
\item{object}{A \link{Candidates} object.}

\item{length_pdf}{Named list keyed by selected reference \code{model_id}.}
}
\value{
Updated \link{Candidates} object.
}
\description{
Set user-supplied reference length PDFs
}
\examples{
\dontrun{
candidates <- set_reference_length_pdf(
  candidates,
  length_pdf = list("12" = c(10, 11, 12, 12, 13))
)
}

}
