% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates_s7.R
\name{show.Candidates}
\alias{show.Candidates}
\title{Show a \code{Candidates}}
\arguments{
\item{object}{A \link{Candidates} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{Candidates}
}
\keyword{internal}
