% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurer_s7.R
\name{show.Configurer}
\alias{show.Configurer}
\title{Show a \code{Configurer}}
\arguments{
\item{object}{A \link{Configurer} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{Configurer}
}
\keyword{internal}
