% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjurer_s7.R
\name{show.Conjurer}
\alias{show.Conjurer}
\title{Show a \code{Conjurer}}
\arguments{
\item{object}{A \link{Conjurer} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{Conjurer}
}
\keyword{internal}
