% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learner_s7.R
\name{show.PolicyLearner}
\alias{show.PolicyLearner}
\title{Show a \code{PolicyLearner}}
\arguments{
\item{object}{A \link{PolicyLearner} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{PolicyLearner}
}
\keyword{internal}
