% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{show.PolicyPredictions}
\alias{show.PolicyPredictions}
\title{Show a \code{PolicyPredictions}}
\arguments{
\item{object}{A \link{PolicyPredictions} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{PolicyPredictions}
}
\keyword{internal}
