% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_selector_s7.R
\name{show.PolicySelector}
\alias{show.PolicySelector}
\title{Show a \code{PolicySelector}}
\arguments{
\item{object}{A \link{PolicySelector} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{PolicySelector}
}
\keyword{internal}
