% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_simulator_s7.R
\name{show.PolicySimulator}
\alias{show.PolicySimulator}
\title{Show a \code{PolicySimulator}}
\arguments{
\item{object}{A \link{PolicySimulator} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{PolicySimulator}
}
\keyword{internal}
