% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{show.Referee}
\alias{show.Referee}
\title{Show a \code{Referee}}
\arguments{
\item{object}{A \link{Referee} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{Referee}
}
\keyword{internal}
