% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{show.Scorecard}
\alias{show.Scorecard}
\title{Show a \code{Scorecard}}
\arguments{
\item{object}{A \link{Scorecard} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{Scorecard}
}
\keyword{internal}
