% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_config_s7.R
\name{show.WorkflowConfig}
\alias{show.WorkflowConfig}
\title{Show a \code{WorkflowConfig}}
\arguments{
\item{object}{A \link{WorkflowConfig} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show a \code{WorkflowConfig}
}
\keyword{internal}
