% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_tuning_cache_version}
\alias{similarity_tuning_cache_version}
\title{Report the current similarity-tuning cache version}
\usage{
similarity_tuning_cache_version()
}
\value{
Integer scalar cache version.
}
\description{
Report the current similarity-tuning cache version
}
\keyword{internal}
