% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{smooth_ts_calibration}
\alias{smooth_ts_calibration}
\title{Smooth TS calibration curves}
\usage{
smooth_ts_calibration(ts_cal)
}
\arguments{
\item{ts_cal}{Policy-by-relative-length calibration table.}
}
\value{
A tibble.
}
\description{
Applies spline smoothing to policy-by-relative-length calibration summaries.
}
