% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_branches}
\alias{standardize_branches}
\title{Standardize policy branch filters from a table}
\usage{
standardize_branches(data, branch_col = "equation_branch_filter")
}
\arguments{
\item{data}{Data frame or tibble.}

\item{branch_col}{Branch-filter column name.}
}
\value{
Character vector.
}
\description{
Standardize policy branch filters from a table
}
\keyword{internal}
