% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{summarize_key_missing.Candidates}
\alias{summarize_key_missing.Candidates}
\title{Summarize key-field missingness for a \link{Candidates} object}
\description{
Summarize key-field missingness for a \link{Candidates} object
}
