% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_missing_gate.Candidates}
\alias{summarize_missing_gate.Candidates}
\title{Summarize missingness gate outcomes for a \link{Candidates} object}
\description{
Summarize missingness gate outcomes for a \link{Candidates} object
}
