% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{summarize_tuning_stability}
\alias{summarize_tuning_stability}
\title{Summarize tuning stability across resamples}
\usage{
summarize_tuning_stability(tune_obj)
}
\arguments{
\item{tune_obj}{Result returned by \code{\link[=tune_similarity_matrix]{tune_similarity_matrix()}}.}
}
\value{
A tibble.
}
\description{
Summarize tuning stability across resamples
}
\keyword{internal}
