% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{ts_at_length}
\alias{ts_at_length}
\title{Compute TS at length}
\usage{
ts_at_length(slope, intercept, lengths_cm)
}
\arguments{
\item{slope}{Numeric slope value.}

\item{intercept}{Numeric intercept value.}

\item{lengths_cm}{Numeric length vector in centimeters.}
}
\value{
Numeric vector of TS values.
}
\description{
Compute TS at length
}
\keyword{internal}
