% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_benchmark.R
\name{ts_from_length}
\alias{ts_from_length}
\title{Compute TS-length values}
\usage{
ts_from_length(slope, intercept, length_cm)
}
\arguments{
\item{slope}{Numeric slope vector.}

\item{intercept}{Numeric intercept vector.}

\item{length_cm}{Numeric length vector.}
}
\value{
Numeric vector.
}
\description{
Compute TS-length values
}
\keyword{internal}
