% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_admissibility_section}
\alias{validate_admissibility_section}
\title{Validate the admissibility section}
\usage{
validate_admissibility_section(admissibility_section, registry_path = NULL)
}
\arguments{
\item{admissibility_section}{Config \code{admissibility} section.}

\item{registry_path}{Optional trait-registry path.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the admissibility section
}
\keyword{internal}
