% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config_paths}
\alias{validate_config_paths}
\title{Validate config paths}
\usage{
validate_config_paths(paths_section)
}
\arguments{
\item{paths_section}{Config \code{paths} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate config paths
}
\keyword{internal}
