% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_execution_flags}
\alias{validate_execution_flags}
\title{Validate execution flags}
\usage{
validate_execution_flags(execution_section, paths_section)
}
\arguments{
\item{execution_section}{Config \code{execution} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate execution flags
}
\keyword{internal}
