% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_metalearner_method_settings}
\alias{validate_metalearner_method_settings}
\title{Validate learner-specific method settings}
\usage{
validate_metalearner_method_settings(method_settings)
}
\arguments{
\item{method_settings}{Metalearner \code{method_settings} list.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate learner-specific method settings
}
\keyword{internal}
