% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_metalearner_section}
\alias{validate_metalearner_section}
\title{Validate the metalearner section}
\usage{
validate_metalearner_section(metalearner_section)
}
\arguments{
\item{metalearner_section}{Workflow \code{metalearner} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the metalearner section
}
\keyword{internal}
