% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_policy_list_section}
\alias{validate_policy_list_section}
\title{Validate the policy-list section}
\usage{
validate_policy_list_section(policies_section, policy_path = NULL)
}
\arguments{
\item{policies_section}{Config \code{policies} section.}

\item{policy_path}{Optional policy-registry path.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the policy-list section
}
\keyword{internal}
