% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_post_selection_conformal_section}
\alias{validate_post_selection_conformal_section}
\title{Validate the post-selection conformal section}
\usage{
validate_post_selection_conformal_section(conformal_section)
}
\arguments{
\item{conformal_section}{Config \code{post_selection_conformal} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the post-selection conformal section
}
\keyword{internal}
