% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_workflow_flags}
\alias{validate_workflow_flags}
\title{Validate workflow flags}
\usage{
validate_workflow_flags(workflow_section, paths_section)
}
\arguments{
\item{workflow_section}{Workflow \code{workflow} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate workflow flags
}
\keyword{internal}
