% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_workflow_paths}
\alias{validate_workflow_paths}
\title{Validate workflow paths}
\usage{
validate_workflow_paths(paths_section)
}
\arguments{
\item{paths_section}{Workflow \code{paths} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate workflow paths
}
\keyword{internal}
