% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility.R
\name{weight_anchor_models}
\alias{weight_anchor_models}
\title{Combine anchor distances to final weights}
\usage{
weight_anchor_models(model_eval, sim_obj, config)
}
\arguments{
\item{model_eval}{Anchor scoring table with kernel terms.}

\item{sim_obj}{Prepared similarity object.}

\item{config}{Anchor config list.}
}
\value{
A tibble.
}
\description{
Combine anchor distances to final weights
}
\keyword{internal}
