% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{weighted_interval_quantiles}
\alias{weighted_interval_quantiles}
\title{Compute multiple weighted central interval quantiles}
\usage{
weighted_interval_quantiles(x, w, widths = c(0.8, 0.9, 0.95, 0.99))
}
\arguments{
\item{x}{Numeric values.}

\item{w}{Numeric weights.}

\item{widths}{Interval widths.}
}
\value{
Named numeric vector.
}
\description{
Compute multiple weighted central interval quantiles
}
\keyword{internal}
