% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{Compute weighted quantiles}
\usage{
weighted_quantile(x, w, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{x}{Numeric values.}

\item{w}{Numeric weights.}

\item{probs}{Numeric probability vector.}
}
\value{
Numeric vector.
}
\description{
Compute weighted quantiles
}
\keyword{internal}
