% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{weighted_upper_quantiles}
\alias{weighted_upper_quantiles}
\title{Compute multiple weighted upper quantiles}
\usage{
weighted_upper_quantiles(x, w, levels = c(0.8, 0.9, 0.95, 0.99))
}
\arguments{
\item{x}{Numeric values.}

\item{w}{Numeric weights.}

\item{levels}{Upper-tail quantile levels.}
}
\value{
Named numeric vector.
}
\description{
Compute multiple weighted upper quantiles
}
\keyword{internal}
