% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{workflow_data}
\alias{workflow_data}
\title{Resolve workflow-config data}
\usage{
workflow_data(config)
}
\arguments{
\item{config}{Workflow-config input.}
}
\value{
A workflow-config list.
}
\description{
Resolve workflow-config data
}
\keyword{internal}
