% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{workflow_legacy_names}
\alias{workflow_legacy_names}
\title{Return disallowed legacy workflow names}
\usage{
workflow_legacy_names()
}
\value{
Named list of legacy-to-current field-name mappings by section.
}
\description{
Return disallowed legacy workflow names
}
\keyword{internal}
