% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{workflow_option_values}
\alias{workflow_option_values}
\title{Build workflow option values}
\usage{
workflow_option_values(config)
}
\arguments{
\item{config}{Normalized workflow-config list.}
}
\value{
Named list of option values.
}
\description{
Returns a named list of workflow options suitable for \code{options(...)}.
}
