% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{workflow_progress}
\alias{workflow_progress}
\title{Emit one workflow progress message}
\usage{
workflow_progress(progress, ...)
}
\arguments{
\item{progress}{Logical scalar.}

\item{...}{Message fragments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Emit one workflow progress message
}
\keyword{internal}
