% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark a staged staged object}
\usage{
benchmark(object, ...)
}
\arguments{
\item{object}{A staged staged object such as a \link{PolicySelector}.}

\item{...}{Method-specific arguments.}
}
\value{
An updated staged staged object.
}
\description{
Dispatches to the appropriate benchmark method for the supplied staged
staged object.
}
