% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{bind_sensitivity_data.default}
\alias{bind_sensitivity_data.default}
\title{Bind scenario benchmark tables from a named scenario-result map}
\arguments{
\item{sensitivity_map}{Named scenario benchmark map.}
}
\description{
Bind scenario benchmark tables from a named scenario-result map
}
