% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{build_species_coverage}
\alias{build_species_coverage}
\title{Build a species-block coverage table}
\usage{
build_species_coverage(pseudo_sum, ...)
}
\arguments{
\item{pseudo_sum}{Pseudo-anchor conformal summary list or a
\link{PolicySelector} object.}

\item{...}{Method-specific arguments.}
}
\value{
A tibble.
}
\description{
Build a species-block coverage table
}
\examples{
\dontrun{
build_species_coverage(selector)
}

}
