% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calibrate_uncertainty}
\alias{calibrate_uncertainty}
\title{Calibrate uncertainty}
\usage{
calibrate_uncertainty(object, ...)
}
\arguments{
\item{object}{A package object such as a \link{PolicySelector} or
\link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
The updated object.
}
\description{
Dispatches to the appropriate uncertainty-calibration method for the
supplied object.
}
