% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidate_models}
\alias{candidate_models}
\title{Preprocessed candidate models table}
\format{
\subsection{\code{candidate_models}}{
}
}
\usage{
data(candidate_models)
}
\description{
A tibble containing species- and study-level metadata traits used for
quantifying model similarities.
}
\keyword{datasets}
