% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{collect_component_weights}
\alias{collect_component_weights}
\title{Collect per-component tuned multipliers}
\usage{
collect_component_weights(tune_obj, resample_id)
}
\arguments{
\item{tune_obj}{Result returned by \code{\link[=tune_similarity_empirical]{tune_similarity_empirical()}}.}

\item{resample_id}{Integer resample identifier.}
}
\value{
A tibble.
}
\description{
Collect per-component tuned multipliers
}
\keyword{internal}
