% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{construct_sensitivity_table}
\alias{construct_sensitivity_table}
\title{Construct a scenario manifest}
\usage{
construct_sensitivity_table(scenario_specifications, ...)
}
\arguments{
\item{scenario_specifications}{Named scenario-specification list.}

\item{...}{Method-specific arguments.}
}
\value{
A tibble.
}
\description{
Summarizes the scenario specifications and the corresponding benchmark
outputs into one scenario-level manifest table.
}
\examples{
\dontrun{
simulator <- as_policysimulator(selector)
simulator <- simulate(simulator)
construct_sensitivity_table(simulator)
}

}
