% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{construct_species_coverage}
\alias{construct_species_coverage}
\title{Build a species-block coverage table}
\usage{
construct_species_coverage(coverage_summary, ...)
}
\arguments{
\item{coverage_summary}{Pseudo-anchor conformal summary list or a
\link{PolicySelector} object.}

\item{...}{Method-specific arguments.}
}
\value{
A tibble.
}
\description{
Build a species-block coverage table
}
\examples{
\dontrun{
construct_species_coverage(selector)
}

}
