% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{crossfit}
\alias{crossfit}
\title{Cross-fit a policy learner}
\usage{
crossfit(object, ...)
}
\arguments{
\item{object}{A \link{PolicyLearner} object.}

\item{...}{Method-specific arguments.}
}
\value{
The updated \link{PolicyLearner} object.
}
\description{
Estimates out-of-fold policy-performance predictions for a learner object.
}
\examples{
\dontrun{
learner <- as_policylearner(selector)
learner <- crossfit(learner)
}

}
