% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{crossfit}
\alias{crossfit}
\title{Cross-fit a staged staged object}
\usage{
crossfit(object, ...)
}
\arguments{
\item{object}{A staged staged object such as a \link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
An updated staged staged object.
}
\description{
Cross-fits the supplied staged staged object using the method associated
with its class.
}
\examples{
\dontrun{
learner <- as_policy_learner(selector)
learner <- crossfit(learner)
}

}
