% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{equal_start_weights}
\alias{equal_start_weights}
\title{Replace starting trait weights with equal weights}
\usage{
equal_start_weights(trait_spec)
}
\arguments{
\item{trait_spec}{Trait specification.}
}
\value{
Named numeric vector or original value.
}
\description{
Replace starting trait weights with equal weights
}
\keyword{internal}
