% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_areas}
\alias{fao_areas}
\title{FAO area polygons}
\format{
\subsection{\code{fao_areas}}{
}
}
\usage{
data(fao_areas)
}
\description{
A tibble containing FAO major area metadat and polygons.
}
\keyword{datasets}
