% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit}
\alias{fit}
\title{Fit a model object}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A package object such as a \link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
The updated object.
}
\description{
Fits the supplied object using the method associated with its class.
}
\examples{
\dontrun{
learner <- as_policylearner(selector)
learner <- fit(learner)
}

}
