% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit}
\alias{fit}
\title{Fit a staged staged object}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A staged staged object such as a \link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
An updated staged staged object.
}
\description{
Fits the supplied staged staged object using the method associated with its
class.
}
\examples{
\dontrun{
learner <- as_policylearner(selector)
learner <- fit(learner)
}

}
