% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{fit}
\alias{fit}
\title{Fit a staged workflow object}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A staged workflow object such as a \link{PolicyLearner}.}

\item{...}{Method-specific arguments.}
}
\value{
An updated staged workflow object.
}
\description{
Fits the supplied staged workflow object using the method associated with its
class.
}
\examples{
\dontrun{
learner <- as_policy_learner(selector)
learner <- fit(learner)
}

}
