% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{forge_distances}
\alias{forge_distances}
\title{Forge a learned distance matrix from an Alchemist}
\usage{
forge_distances(object, ...)
}
\arguments{
\item{object}{An \link{Alchemist} object.}

\item{...}{Method-specific arguments.}
}
\value{
An updated \link{Alchemist} object with \verb{@distance_matrix} and \verb{@learner}
populated.
}
\description{
Fits a Super Learner to pairwise acoustic distances and stores the resulting
N x N learned distance matrix on the object.
}
