% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{forge_distances}
\alias{forge_distances}
\title{Learn an Alchemist distance matrix}
\usage{
forge_distances(object, ...)
}
\arguments{
\item{object}{An \link{Alchemist} object.}

\item{...}{Method-specific arguments.}
}
\value{
An updated \link{Alchemist} object containing the learned distance
bundle and fitted learner state.
}
\description{
Fits a Super Learner to pairwise acoustic distances and stores the resulting
model-by-model learned distance matrix on the object.
}
