% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_metalearner_section}
\alias{normalize_metalearner_section}
\title{Normalize the metalearner section}
\usage{
normalize_metalearner_section(metalearner_section)
}
\arguments{
\item{metalearner_section}{Workflow \code{metalearner} section.}
}
\value{
Normalized \code{metalearner} section.
}
\description{
Normalize the metalearner section
}
\keyword{internal}
