% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_trait_map}
\alias{normalize_trait_map}
\title{Normalize one config trait map}
\usage{
normalize_trait_map(trait_map)
}
\arguments{
\item{trait_map}{Optional named trait-weight map.}
}
\value{
Named numeric vector.
}
\description{
Normalize one config trait map
}
\keyword{internal}
