% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{normalize_trait_map}
\alias{normalize_trait_map}
\title{Normalize one workflow trait map}
\usage{
normalize_trait_map(trait_map, trait_cols, trait_weights)
}
\arguments{
\item{trait_map}{Optional named trait-weight map.}

\item{trait_cols}{Optional character vector of trait names.}

\item{trait_weights}{Optional named trait-weight map.}
}
\value{
Named numeric vector.
}
\description{
Normalize one workflow trait map
}
\keyword{internal}
