% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_area_distribution}
\alias{plot_area_distribution}
\title{Plot FAO study distribution map}
\usage{
plot_area_distribution(model_data, count_type = "studies")
}
\arguments{
\item{count_type}{The type of count to distribute across FAO major regions.
This can either be 'studies' for the number of studies, or 'models' for the
number of models.}

\item{source_data}{Model metadata table.}
}
\value{
A ggplot object.
}
\description{
Plot FAO study distribution map
}
