% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ordination_cluster_hulls}
\alias{plot_ordination_cluster_hulls}
\title{Plot ordination clusters with hulls}
\usage{
plot_ordination_cluster_hulls(
  points_tbl,
  hull_tbl,
  cluster_col = "policy_cluster_id",
  reference_col = "is_reference",
  label_col = "species_name",
  title = "NMDS Clusters with Hulls"
)
}
\arguments{
\item{points_tbl}{Ordination point table.}

\item{hull_tbl}{Cluster-hull table.}

\item{cluster_col}{Cluster-label column.}

\item{reference_col}{Reference-flag column.}

\item{label_col}{Point-label column.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot ordination clusters with hulls
}
