% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_ts_bands}
\alias{plot_ts_bands}
\title{Plot TS conformal ribbons}
\usage{
plot_ts_bands(
  band_tbl,
  curve_tbl,
  anchor_label,
  policy_label,
  show_top_candidate = FALSE
)
}
\arguments{
\item{band_tbl}{TS ribbon-band table.}

\item{curve_tbl}{TS summary curve table.}

\item{anchor_label}{Anchor label used in the title.}

\item{policy_label}{Selected policy label used in the subtitle.}
}
\value{
A ggplot object.
}
\description{
Plot TS conformal ribbons
}
