% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{preview_values}
\alias{preview_values}
\title{Collapse values for compact display}
\usage{
preview_values(values, limit = 5L)
}
\arguments{
\item{values}{Character vector.}

\item{limit}{Maximum number of values to display.}
}
\value{
Character scalar.
}
\description{
Collapse values for compact display
}
\keyword{internal}
