% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee.R
\name{referee_rebuild}
\alias{referee_rebuild}
\title{Rebuild a \code{Referee}}
\usage{
referee_rebuild(
  object,
  selector = NULL,
  learner = NULL,
  predictions = NULL,
  config = NULL,
  scorecard = NULL
)
}
\arguments{
\item{object}{A \link{Referee} object.}

\item{selector}{Optional replacement \link{PolicySelector}.}

\item{learner}{Optional replacement \link{PolicyLearner}.}

\item{predictions}{Optional replacement \link{PolicyPredictions}.}

\item{config}{Optional replacement config list.}

\item{scorecard}{Optional replacement \link{Scorecard}.}
}
\value{
A \code{Referee} object.
}
\description{
Reconstructs a \link{Referee} object, optionally replacing one or more of its
component objects.
}
