% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee_s7.R
\name{referee_rebuild}
\alias{referee_rebuild}
\title{Test whether an object is a \code{Referee} instance}
\usage{
referee_rebuild(
  object,
  selector = object@selector,
  learner = object@learner,
  predictions = object@predictions,
  config = object@config,
  scorecard = object@scorecard
)
}
\arguments{
\item{object}{A \link{Referee} object.}

\item{selector}{Optional replacement \link{PolicySelector}.}

\item{learner}{Optional replacement \link{PolicyLearner}.}

\item{predictions}{Optional replacement \link{PolicyPredictions}.}

\item{config}{Optional replacement config list.}

\item{scorecard}{Optional replacement \link{Scorecard}.}
}
\value{
A \code{Referee} object.
}
\description{
Rebuild a \code{Referee}
}
