% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee.R
\name{Scorecard-class}
\alias{Scorecard-class}
\alias{Scorecard}
\title{Typed recommendation scorecard}
\description{
Stores the post-prediction recommendation tables and diagnostics built by
\code{\link[=predict]{predict()}} on a \link{Referee}.
}
\examples{
\dontrun{
scorecard <- predict(as_referee(selector, predictions = predictions))
scorecard
scorecard
}

}
