% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_missing_gate.default}
\alias{summarize_missing_gate.default}
\title{Summarize missingness gate outcomes for admissibility summary tables}
\arguments{
\item{adm_tbl}{Admissibility summary table.}
}
\description{
Summarize missingness gate outcomes for admissibility summary tables
}
