% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{summarize_missing_gate}
\alias{summarize_missing_gate}
\title{Summarize missingness gate outcomes}
\usage{
summarize_missing_gate(adm_tbl, ...)
}
\arguments{
\item{adm_tbl}{Admissibility summary table.}

\item{...}{Method-specific arguments.}
}
\value{
A tibble.
}
\description{
Summarize missingness gate outcomes
}
\examples{
\dontrun{
summarize_missing_gate(selector)
}

}
