% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_missing_gate}
\alias{summarize_missing_gate}
\title{Summarize missingness gate outcomes}
\usage{
summarize_missing_gate(adm_tbl)
}
\arguments{
\item{adm_tbl}{Admissibility summary table.}
}
\value{
A tibble.
}
\description{
Summarize missingness gate outcomes
}
