% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{tsb_message}
\alias{tsb_message}
\title{Emit a timestamped workflow message}
\usage{
tsb_message(..., timestamp = TRUE, appendLF = TRUE)
}
\arguments{
\item{...}{Message components.}

\item{timestamp}{Boolean that dictates whether to prepend with timestamp.}

\item{appendLF}{Passed to \code{base::message()}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Emit a timestamped workflow message
}
\keyword{internal}
