% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_benchmark_section}
\alias{validate_benchmark_section}
\title{Validate the benchmark section}
\usage{
validate_benchmark_section(benchmark_section)
}
\arguments{
\item{benchmark_section}{Config \code{benchmark} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the benchmark section
}
\keyword{internal}
