% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_cache_section}
\alias{validate_cache_section}
\title{Validate the cache section}
\usage{
validate_cache_section(cache_section)
}
\arguments{
\item{cache_section}{Workflow \code{cache} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the cache section
}
\keyword{internal}
