% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_selection_section}
\alias{validate_selection_section}
\title{Validate the selection section}
\usage{
validate_selection_section(selection_section)
}
\arguments{
\item{selection_section}{Workflow \code{selection} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the selection section
}
\keyword{internal}
