% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_stage_section}
\alias{validate_stage_section}
\title{Validate a simple stage section}
\usage{
validate_stage_section(stage_section, section_name, worker_field = NULL)
}
\arguments{
\item{stage_section}{Workflow stage section.}

\item{section_name}{Section name for error messages.}

\item{worker_field}{Optional worker-field name.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate a simple stage section
}
\keyword{internal}
