% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_tuning_section}
\alias{validate_tuning_section}
\title{Validate the tuning section}
\usage{
validate_tuning_section(tuning_section)
}
\arguments{
\item{tuning_section}{Workflow \code{tuning} section.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Validate the tuning section
}
\keyword{internal}
